# small optimization to IP quantization