<!--
* Contributors are encouraged to read our [CONTRIBUTING](/CONTRIBUTING.md) documentation.
* Both Contributor and Reviewer Checklists are available at https://datadoghq.dev/datadog-agent/guidelines/contributing/#pull-requests.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's change.
* Please fill the below sections if possible with relevant information or links.
-->
### What does this PR do?

This PR improves the performance of the IP quantization operation by skipping the regex call in most cases: when the hints are not present in the given text. [Example profile](https://ddstaging.datadoghq.com/profiling/explorer?query=service%3Atrace-agent&agg_m=count&agg_m_source=base&agg_t=count&event=AgAAAZLTzNMqkdMaLQAAAAAAAAAYAAAAAEFaTFR6TlNvQUFDVzBTTFZSVjN2ZGdBQQAAACQAAAAAMDE5MmQzY2MtZjc0My00ZjQzLWExMjktMjU3YTVlY2M1NDUz&my_code=disabled&profileId=AZLTzNSoAACW0SLVRV3vdgAA&refresh_mode=paused&viz=stream&from_ts=1730126668534&to_ts=1730130268534&live=false) showing that this code path was quite present.

Benchmark (CPU only, allocations are the same):
```
goos: darwin
goarch: arm64
pkg: github.com/DataDog/datadog-agent/pkg/obfuscate
cpu: Apple M1 Max
                            │  before.txt   │              after.txt              │
                            │    sec/op     │   sec/op     vs base                │
SplitPrefix/matching-10         342.6n ± 9%   351.9n ± 2%        ~ (p=0.078 n=10)
SplitPrefix/not_matching-10   1378.00n ± 1%   37.22n ± 1%  -97.30% (p=0.000 n=10)
geomean                         687.1n        114.4n       -83.34%
```


### Motivation

### Describe how to test/QA your changes

### Possible Drawbacks / Trade-offs

### Additional Notes
<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the proposed
  version was chosen.
-->